Fix Elements Not Loading On AJAX Checkout
WooCommerce can load the `#stripe-card-element` dynamically in fragments via AJAX (`.woocommerce-checkout-payment` fragment in response to `update_order_review` action). If we check for `#stripe-card-element` on page load, Stripe Elements are not mounted when the fragments are loaded later. This change pushes the check for `#stripe-card-element` into the `'updated_checkout` event handler, at which point it has been made available via the fragment.