Creates a complete postcode when calculating shipping
This addresses a specific issue I had where shipping would be allowed to partial postcode areas like N3, SW11, etc. but woocommerce could not work out the specifics without complete postcodes (ApplePay would only return the first part). As a fix for this issue when a postcode that is too short is received we can fake a complete postcode by adding "1AA" as the postcode does not need to be valid for woocommerce to correctly calculate shipping.