Always return the last charge made to a PaymentIntent, instead of the first one.
After calling `/payment_intents/xxxxx/capture`, the returned response is guaranteed to have at least 1 `Charge` object in its list. But, it's possible for a Payment intent to have multiple charges, for example for partial payments or if there have been retries on the same Intent object. It's safer to return the last Charge object instead of the first.