Make the publishable key fields be plain text instead of "password" (#915)
Fixes #895

The test/live publishable keys aren't "secrets", they are injected into the DOM on the checkout in plain text. There's no need for those fields to be `type=password`, it just makes it more annoying to setup.

To test: Go to the Stripe payment method settings, see that the `Test Publishable Key` / `Live Publishable Key` input fields are displayed in plain text.