Remove the deprecated "betas" parameter so Stripe.js doesn't log a warning (#1001)
The `betas` parameter needed to be passed to the `stripe.js` initialization code so Payment Intents would work. Now that Payment Intents have been released out of Beta for a while, that `betas` parameter is no longer needed. In fact, it doesn't even appear in Stripe's documentation anymore.

This is a simple change, I just removed the `betas` parameter and run `npm run build`.

Fixes #998