Fix validation of subscription payment meta
When the subscription is updated, the function
`validate_subscription_payment_meta` is called to validate all the
fields, regardless which actual action the user wants to perform.
It's called by the hook woocommerce_subscription_validate_payment_meta.

If validation is performed during subscription renewal it's allowed to
have empty stripe customer id becausee it will be added when payment
is processed if required.