Add missing customer ID to subscriptions before processing payment (#1107)
* Add missing customer ID to subscriptions before processing payment

Address subscriptions that might not have a Stripe customer ID when renewing. This could happen if the customer updated their recurring payment method (see issue #1072). PR #1106 proposed a fix for new recurring payment updates. This one aims to address existing subscriptions that cannot be renewed and have to be manually updated by merchants/customers.

* Fix validation of subscription payment meta

When the subscription is updated, the function
`validate_subscription_payment_meta` is called to validate all the
fields, regardless of which actual action the user wants to perform.
It's called by the hook woocommerce_subscription_validate_payment_meta.

If validation is performed during subscription renewal it's allowed to
have empty stripe customer id because it will be added when payment
is processed if required.

Co-authored-by: Vasily Belolapotkov <belolapotkov.v@gmail.com>