Improve branded Google Pay button user agent detection (#1216)
The current implementation for Chrome detection is matching some Chromium-based browsers.
Achieving 100% coverage of existing browsers and maintaining the checks is a huge effort,
but we should do it for the most used ones. This commit adds checks to prevent the GPay
button from rendering in Edge, Opera and Brave.