Fix incorrect order summary on single product page (#1239)
When using the payment request buttons (apple pay/google pay) on a
product's page the order summary would initially show the product's
name (itemized view). However, after changing either the shipping
address or shipping method the order summary would only show the
subtotal.

This fixes that behavior by passing a variable from the frontend -
`is_product_page` - to the backend so we can make a decision about
whether to use an itemized view (product's name in this case) or show
the subtotal.