Remove phone number requirement with Apple Pay
This commit removes the hard requirement for customers to provide a
phone number when checking out with Apple Pay.

There is now an option in the Stripe settings that allows you to
re-enable the phone number requirement. This is a suggestion and is up
for discussion. This may be better suited as a filter, since most
merchants probably don't want the phone number requirement anyway.