Rely on the WooCommerce Customizer phone settings
This is different from the previous implementation that added an option
to the Stripe plugin settings page.

The (slight) disadvantage to this implementation is that the phone
number is required by default, which might not be something we want.
The upside, however, is that we're fully aligned with the WC core
settings.