Use AJAX API for successful PRB checkout
This commit is pretty big, but I'm not really sure how to split it into
smaller chunks. For the PRB to function correctly all 3 handlers must be
implemented; shipping options change, shipping address change, and the
payment handler.

That said, 3DS payment function correctly with this commit, but the
Blocks checkout does not. As an example of things the blocks checkout
doesn't currently do:

* Show error messages under certain conditions, e.g. failing the 3DS
  authentication.
* The checkout isn't disabled during payment processing.

Further integration with the Blocks API is needed, but this is a good
first step I think.