Use AJAX APIs to checkout with PRBs
We now control the PRBs entirely through AJAX APIs. Errors are reported
correctly and payments are fully functional.

The PRBs and the Blocks are still disconnected though, so any update
made in one (e.g. changing the shipping address) is not reflected in the
other.

For now, the focus is on cleaning up the code, what follows is
implementing the integration with the Blocks.