Add comment explaning why do we need to remove the call to `getSelectedProductData`
The response from `ajax_get_selected_product_data` comes incomplete if there’s nothing in the Cart.

If the Cart is empty and the user clicks the PR button the response will be almost empty. `getSelectedProductData` should return a valid response so when used it can be user to update `paymentRequest` properly. Now, this is useless given the next approach we are implementing where we get most of the data for the response object from the Cart. On the other hand, we shouldn’t add to cart the product when changing the quantity because that’s not how it was done before.

Leaving it as is for now until we decide what approach to follow here.