Use shared configuration for shortcode- and block-based checkout flows (#1585)
The JS configurations are now generated by isolated functions
so we can use the generated arrays in multiple places; specifically
for both the Block-based and Shortcode-based checkout flows.

Additionally this PR makes it so that only relevant JS code is
loaded on checkout pages. If there's a block we load the Block
JS, if there's a Shortcode we load the Shortcode JS.
To that end we made a slight change in how we decide when
the Shortcode JS is loaded. Instead of not loading the
shortcode JS when a Block is present on the current page,
we now load the shortcode JS when a cart or checkout
shortcode is present on the page.