Allow custom status when payment processing fails
Made possible via 'wc_stripe_allowed_payment_processing_statuses'
filter.

Before this commit a merchant would not be alerted if a payment fails
when the order has a custom status. By adding the custom status to the
list of allowed statuses through the aforementioned filter the merchant
is alerted when a payment fails.

Note that this is for unusual failures such as the Stripe API being
unavailable. This does not cover "expected" errors such as when a
customer doesn't have enough funds in their account.