Fix payments with custom order status made via payment or setup intents (#1657)
* Add allowed_payment_processing_statuses filter

This filter is used to control which order statuses allow a payment to
be processed in the context of 3DS payments and the associated webhooks.
By default the only statuses treated as such are 'pending' and 'failed'.

This is mostly useful when merchants assign a custom status to an order
before payment processing begins. Before this commit a custom order
status would capture the payment, but the order status wasn't updated.

See PR #1499 for more information.

Example usage:

    public function allow_reserved_order_status( $statuses ) {
        // This assumes a 'reserved' order status slug already exists.
        if ( ! in_array( 'reserved', $statuses ) ) {
            $statuses[] = 'reserved';
        }
        return $statuses;
    }
    add_filter(
        'wc_stripe_allowed_payment_processing_statuses',
        'allow_reserved_order_status'
    );


* Allow custom status when payment processing fails

Made possible via 'wc_stripe_allowed_payment_processing_statuses'
filter.

Before this commit a merchant would not be alerted if a payment fails
when the order has a custom status. By adding the custom status to the
list of allowed statuses through the aforementioned filter the merchant
is alerted when a payment fails.

Note that this is for unusual failures such as the Stripe API being
unavailable. This does not cover "expected" errors such as when a
customer doesn't have enough funds in their account.