Add separate function for Blocks config
The getStripeServerData() function would return the Classic UPE config
even if the Checkout block was being used on the current page. This
caused an issue in the Checkout Block integration where the Stripe
payment method would not be made available.

The fix was to create a new getStripeServerData() function specific to
the Blocks integration and use that everywhere in client/blocks.