Adds statement previews for full and shortened bank statement text. (#1871)
This creates a new StatementPreview component that accepts icon, title, text, and className props. The icon should be a string that matches a pre-defined property name of an icon SVG in the component. Current options for this are `creditCard` and `bank`. The title is a string of text shown at the top of the preview and describes the preview. text is the text that will show under the "Transaction" heading on the preview. And className will be added to the outer div of the preview card.

The "amount" shown on the card will respect the shop's currency settings for currency symbol, currency symbol position, decimal separator, and precision.