Adds "Disable Stripe" confirmation modal (#1932)
This PR adds a "disable Stripe" confirmation modal when a user attempts to disable the Stripe payment method when UPE is enabled and they have 1 or more payment methods added or express checkouts are enabled. If the user cancels the modal, the gateway remains enabled, if they confirm it the gateway is disabled.