Add upe integration for the Checkout Block (#1825)
* Fix issue where ajax endpoint was duplicated

When array_merge_recursive is used properties with the same name are
added to an array. This made it so that the ajax_url property was
duplicated in the frontend causing API requests to fail.

By using array_replace_recursive instead we fix this issue.

* Render UPE element in the Checkout Block

* Add separate function for Blocks config

The getStripeServerData() function would return the Classic UPE config
even if the Checkout block was being used on the current page. This
caused an issue in the Checkout Block integration where the Stripe
payment method would not be made available.

The fix was to create a new getBlocksConfiguration() function specific to
the Blocks integration and use that everywhere in client/blocks.

* Ensure stripe JS is loaded for UPE support

* Pass configuration into WCStripeAPI

This removes the dependency on the external configuration which will
ultimately make the class more testable as well as being easier across
different use cases going forward.