Update payment request when billing data changes
Applies only to the Cart Block integration.

The PaymentRequestButtonElement doesn't support re-assigning the
paymentRequest prop, and as such we have to update the payment request
instead.

We achieve this by sending another API request for the cart to make sure
all normalizations are correctly applied, and use that data to update
the payment request itself.