UPE: Handle requires_confirmation status for PIs
When a 3DS authorization fails and you immediately try to pay with the
same card the payment intent transitions to a 'requires_confirmation'
status. To fix the issue we make sure the intent is confirmed when in
that state instead of skipping over it like we did before this change.

This change only applies to UPE.