Add filters to disable scripts on product and cart
These filters will only be applied when the PRBs are disabled on the
product and cart pages. This will give merchants the option to not load
the extra JS when they're not using the PRBs on those pages.

Stripe's JS includes some fraud prevention mechanisms, so disabling the
scripts using these filters does increase the risk of successul fraud on
a merchant's store, but that's something merchants will have to decide
on their own.