Hide "Buy Now" button when all variations are out of stock
Before, the "Buy Now" button would appear even if all variations of a
product were out-of-stock. This behavior did not match the one that
existed in the simple product pages.

Now, the "Buy Now" button only appears if there exists at least one
variation that is not out-of-stock. This fixes the mismatching behavior.

Fixes #1580