Remove filter for unloading Stripe JS with blocks
The function used to retrieve the script handles is run at a point where
the block hasn't been loaded, so we can't detect whether we're on a cart
or product page. As such we can't make an informed decisions on whether
or not to load the script.

The script is loaded asynchronously in the blocks checkout though, so
this shouldn't be degradation in performance as it relates to the issue
reported in #2065.