Only load Stripe JS on product and cart page when required (#2110)
Previously we always loaded the Stripe JS on relevant pages through the
main gateway class, but now we have an additional check there to make
sure that doesn't happen.

* Add filters to disable scripts on product and cart

These filters will only be applied when the PRBs are disabled on the
product and cart pages. This will give merchants the option to not load
the extra JS when they're not using the PRBs on those pages.

Stripe's JS includes some fraud prevention mechanisms, so disabling the
scripts using these filters does increase the risk of successul fraud on
a merchant's store, but that's something merchants will have to decide
on their own.