Hide PRB when all variations are out-of-stock (#2128)
* Hide "Buy Now" button when all variations are out of stock

Before, the "Buy Now" button would appear even if all variations of a
product were out-of-stock. This behavior did not match the one that
existed in the simple product pages.

Now, the "Buy Now" button only appears if there exists at least one
variation that is not out-of-stock. This fixes the mismatching behavior.

Fixes #1580

* Hide "Buy Now" button if a given variation is out-of-stock

Before, the "Buy Now" button would be displayed, even if the variation
was out-of-stock. This gave the false impression that the variation
could still be purchased.

Now, the "Buy Now" button is hidden when the user selects a variation
that is out-of-stock.

* Add changelog entry: Hide PRB when variations are out-of-stock

* Hide "variable-subscription" if all variations out-of-stock

Applying the same rule to hide "variable-subscription", like the one
that is applied to "variable" products.

Before, the payment request button (PRB) would be hidden if all
variations of a "variable" product were out of stock but not if all
variations of a "variable-subscription" product. This was inconsistent
behavior.

* Update logic that determines if all variations are out-of-stock

Leveraging an existing method (e.g. 'get_available_variations()') to
retrieve all of the variations and parse out whether there's at least
one "in-stock".

The tradeoff here is that 'get_available_variations()' makes multiple
calls to the DB (through 'wc_get_product($variation_id)') in order to
retrieve the data of each variation. But at the same time, we obtain
more readable, hence, more maintainable code.

* Immediately hide/display PRB, based on whether it's in-stock

Before, the payment request button (PRB) would wait for a backend
request to respond before hiding/displaying the PRB. This created a bit
of an odd behavior.

Now, the PRB is displayed/hidden based on the product's cached-data.

* Rename variables to remove unnecessary suffix (e.g. "_ele")

* Leverage existing event for retrieving variation, on change

Before, the code duplicated existing functionality that retrieved some
selected variation.

Now, the code leverages the existing functionality by subscribing to the
event that gets triggered when some valid variation is selected.