Fix payment method names in payments tab: iDEAL, giropay, & Sofort
Before, when visiting the "Payments" tab of WooCommerce, the title of
the tab would have the payment method names incorrectly capitalized.
This commit fixes that.

Also, the POT extension file's references were updated to point to the
correct file-lines. And, log statements were updated to reflect the
correct capitalization of the payment methods.