Fix payment methods' capitalization (#2218)
* Fix payment method names in sections: iDEAL, giropay, & Sofort

Before, when managing WooCommerce payment methods, the section names of
iDEAL, giropay, & Sofort were incorrectly capitalized. This fixes that
and also updates the POT extension file so that the reference links are
correct.

* Fix payment method names in "Get started with Stripe" description

Before, if one was not connected to Stripe, a card that allowed one to
connect to Stripe would appear. In that card, a description was
displayed where iDEAL and Sofort were not capitalized correctly. This
fixes the capitalization of these payment methods.

* Fix payment method names in admin notices: giropay, iDEAL, & Sofort

Before, if a payment method was enabled but the incorrect currency was
set, a notice would appear in the WooCommerce admin panel. These notices
had payment method names with incorrect capitalization. This has now
been fixed.

* Fix payment method names in customer checkout: iDEAL, giropay, Sofort

Before, when a customer was checking out with either iDEAL, giropay, or
Sofort, an icon would appear but the icon's "alt" text had the payment
method names incorrectly capitalized. This commit fixes that.

* Fix payment method names in payments tab: iDEAL, giropay, & Sofort

Before, when visiting the "Payments" tab of WooCommerce, the title of
the tab would have the payment method names incorrectly capitalized.
This commit fixes that.

Also, the POT extension file's references were updated to point to the
correct file-lines. And, log statements were updated to reflect the
correct capitalization of the payment methods.

* Fix Sofort's capitalization in invalid_sofort_country value

Before, at checkout, if a non-supported country was selected for Sofort,
a message indicating invalid country would appear but the capitalization
was incorrect. This commit fixes that.

* Fix payment method names in WC "Payments" tab: iDEAL, giropay, & Sofort

Before, in WooCommerce (WC) "Payments" tab, the names of some payment
methods were capitalized incorrectly. This commit fixes that.

* Fix Sofort payment method name's capitalization in title and label

Also, fix the corresponding tests.

* Fix payment method names in WC "Payments" tab: iDEAL and Sofort

Also, fix the capitalization of iDEAL in the tests file.

* Fix payment method names in comments: iDEAL, giropay, Sofort

Before, there was an inconsistency between the capitalization of the
payment method names that were displayed to the user versus what was in
the comments. This commit removes that inconsistency.

* Fix payment method names in readme.txt: Sofort, iDEAL, & giropay

This update fixes the capitalization of the payment methods: Sofort,
iDEAL, and giropay in the readme.txt file.

* Fix payment method names in the manual testing instructions

Before, some of the payment methods (iDEAL, giropay, Sofort) were
incorrectly capitalized in the plugin. So, the instructions were also
incorrectly capitalized. But since these payment methods' names have
been updated to reflect correct capitalization, these testing
instructions also need to be updated in order to stay consistent.

* Update changelog: Fixed capitalization for payment method names