Reuse method_title property, as opposed to adding a label property
Because the WC_Stripe_Payment_Gateway class already had a `method_title`
property that has the name of the payment gateway, it was a bit
redundant to add the `label` property. This commit removes the `label`
property and reuses the `method_title` property, instead. This does mean
that the payment method/gateway will read slightly different when UPE is
enabled/disabled.

See the following GH comment for more info:
https://github.com/woocommerce/woocommerce-gateway-stripe/pull/2230#discussion_r767743479