Rollback puppeteer to ^2.0.0
This is the puppeteer version used by WCPay and previously required by @automattic/puppeteer-utils, which is the package we rely on for e2e tests. The version we were using before made the js unit tests fail due to a function from an external package not being mocked as expected.