Account Keys Connection Status (#2273)
* Code cleanup: Remove repetitive code

Before, each store-module was spread into actions, selectors, and
resolvers. This created some repetition and would continue to create
more repetition in the future, as more store-modules would be added.
Now, the store-modules can be added into an array, which auto-handles
the spreading into actions, selectors, and resolvers.

* Code Cleanup: Capitalize the POST rest call

* Add frontend functionality to test account keys connection status

This adds the frontend UI that tests whether the publishable key and the
secret key are valid. It does this by creating a token with the
publishable key, using Stripe's library. Then retrieving the token with
the secret key, using our REST API, which uses Stripe's client library.

* Update tests to include newly added account-keys properties

Account key's store-module's state has newly added properties: isTesting
and isValid. It also includes newly added selectors. This commit
accounts for these new properties.

* Add frontend test to test account-keys connection status functionality

* Add backend functionality to test account keys connection status

This adds the backend functionality to determine if the secret key is
valid. It does this by retrieving a token with a given secret key and
token ID.

* Update changelog: Add ability to test Stripe account keys' validity

* Wrap account-key connection text on mobile

Before, the connection text would be too wide and would make the modal
footer's elements look scattered. This commit forces the connection
text's width to shrink, on mobile.

* Replace expanded REST API URL with NAMESPACE constant

Before, the REST API URL contained the expanded version as opposed to
using the NAMESPACE constant. This made it so that if the URL were to
ever be updated, it would require multiple places to be updated, as well.

* Remove unused code: useGetAccountKeysValidity

* Reset account-key validity upon switching between live/test tabs

Before, when account-keys were tested while on a certain tab and then
the tab was switched, the account-keys' status would remain. This would
be a bit confusing since the form is reset, once the user switches tabs.

* Internationalize text for account-key connection status

* Add notification for when incorrect account keys are entered

Before, test account-keys could be entered in the "live" tab and vice-
versa. This created inconsistent behavior.

* Handle live account-keys status connection on insecure (HTTP) connection

Live account-keys must use HTTPS connection. This commit handles the
case where an insecure (HTTP) connection is used to test a live account-
key.

* Allow Stripe to handle secret-key validation when testing keys

Removing backend validation to remove repetitive code and to allow
Stripe to validate the secret-key, instead.

* Allow connection status test for secret-keys that begin with 'rk'

Before, restrictive secret keys were incorrectly considered invalid
secret keys because they did not begin with the 'sk' prefix. Now, the
'rk' prefix is considered to be valid.