Add onboarding payment gateway setup methods (#1569)
* Add remote payment gateway setup methods

* Add needs_setup method

* Add missing return param

* Filter the onboarding settings depending on types of keys entered

* Use get_option to get latest updated setting values

* Update connection URL method name

* Only toggle test mode based on keys when referer is wc-admin

* PHPCS: Use short array syntax to define arrays

Before, the linter was failing the following PHPCS rule:
- "Short array syntax must be used to define arrays"

* PHPCS: Ignore sanitization for simple query parsing

Before, PHPCS would warn that $_SERVER['HTTP_REFERRER']
was not being sanitized correctly. But, because it is
being used for simple comparisons to check 'page' and
'task', the warning can be surpressed.

* PHPCS: Add translators helper comment & update API details link

Before, the Stripe API details link was referencing the Stripe docs
when it was intended to reference the Stripe account's API details. This
commit fixes that.

Also, fixing a misaligned line of code.

* Add comment to clarify the reason for short-circuiting early

Before, the if-then block wasn't 100% clear as to what was going on.
The added comment helps clarify what's going on in the code, a bit.

* Code cleanup: Wrap long line into multiple lines

Also, replace 'else if' with 'elseif'.

* Enable Stripe payment method after connecting an account

Before, the Stripe payment method was not automatically enabled after
the merchant connected their Stripe account through "Connect" or by
manually entering Stripe account keys.

Now, the Stripe payment method is automatically enabled after the
merchant connects their Stripe account through either approach.

Fixes #2314

* Auto disable Stripe payment method once account keys are removed

Before, if the merchant removed their Stripe account keys, the Stripe
payment method would remain enabled. Now, it is automatically disabled
if the account keys are removed.

* Update changelog: Enable Stripe payment method after connecting

Also: Add onboarding payment gateway setup methods.

* Replacing "deleted" with "disconnected" in comment, for accuracy

The word "deleted" implies that the Stripe account will be deleted when,
in fact, it will be "disconnected". The rewording should prevent any
confusion.

Co-authored-by: Hector Lovo <hector.lovo@automattic.com>