Setup the e2e tests environment using WC e2e boilerplate (#2238)
* Set up the base WC e2e boilerplate

At this point, the docker environment is the default one, without specific plugins or options set up. The tests don't run because puppeteer isn't set.

* Install puppeteer and jest-environment-node

* Add default.json file for the e2e tests

It's copied from WooCommerce, as a sample. We have to update it but it makes the tests run.

* Activate the Stripe plugin in the e2e install

* Create separate specs for merchant and shopper e2e tests flows

This will allow us to better organize the tests.

* Use config.get() to make the test fail

Simply adding a line that will throw an error. For debugging.

* Update the config package to 3.3.3

The previous version was somewhat old and caused unexpected errors when using the package in the e2e tests.

* Add config file for e2e tests

This is the file that's actually retrieved when using the config package within the e2e tests.

* Add a file for setting up the e2e tests flows

It's currently empty, to serve as a base.

* added function to open the upe settings page

* Rename e2e-tests scripts to match the format from WCPay

The previous names were the default ones from the e2e boilerplate. Changing them to match the format from WCPay for consistency.

* Update example test to get rid of the deprecation warning in the logs

This is an example tests that's going to be removed. Just for having a neat run.

* Remove unused test file

* Use a classic 'if' instead of optional chaining within useEffect's callback

The only reason behind this change is that the introduced @woocommerce/e2e-environment package locks Eslint to 6.7.2 which doesn't support optional chaining, making the linting fails in these cases. We could go back to optional chaining once the Eslint version is back to one higher than 7.5.0.

* Make the mocked @wordpress/a11y module virtual

The previous mock stopped working after introducing @woocommerce/e2e-environment. Making it virtual allows mocking its 'speak' function, making the tests work as expected and as they worked before introducing the e2e-environment package.

* Remove clearAllMocks after each test for testing

Giving it a quick test on whether this makes a difference for the tests that fail in GH but pass locally.

* Revert "Remove clearAllMocks after each test for testing"

This reverts commit cee017d0ab76968d5d54a7d8f5d2559e1e4e5a53.

* Clear individual mocks on afterEach

Another attempt to solve the tests passing locally but failing on GH.

* Revert "Clear individual mocks on afterEach"

This reverts commit eb9f40c5a70853c28408828fd9c3dcf130be4ab3.

* Use __esModule for mocking @wordpress/a11y

Is this the reason why the tests fail on GH? Let's see.

* Revert "Use __esModule for mocking @wordpress/a11y"

This reverts commit 9d2aeb475b83189bb0b1da89b0ba74b601666023.

* Remove unused mock of @wordpress/ally

The speak function doesn't seem to be affecting the tests. Also an attempt to figure out why the tests ran from the CI fail.

* Revert "Remove unused mock of @wordpress/ally"

This reverts commit 1cff042f2ae403adedc6b932bb8c770fe58b6fad.

* Rollback puppeteer version to ^2.0.0

This is the puppeteer version used by WCPay and previously required by @automattic/puppeteer-utils, which is the package we rely on for e2e tests. The version we were using before made the js unit tests fail due to a function from an external package not being mocked as expected.

* Lock puppeteer version at 2.1.1

* Remove @woocommerce/api package

For testing.

* Revert "Remove @woocommerce/api package"

This reverts commit e2907cfcaf844f1f4c02213316569cd83f63f6b3.

* Revert "Make the mocked @wordpress/a11y module virtual"

This reverts commit 0af0c20153ada330334e017fcb6116d921cb0e52.

* Update package-lock.json

* Update the package-lock.json file with the one from develop

There's an issue with the file that's built from scratch. This happens on the develop branch as well. Let's build things on top of the develop file for things to work and check that issue out separately.

* Uninstall config, puppeteer, and jest-environment-node packages

These are installed as dependencies of other packages.

* Revert "Use a classic 'if' instead of optional chaining within useEffect's callback"

This reverts commit 595f0550572cd0681674245d820c851e6a7144d8.

* Comment out e2e path to a shopper directory that doesn't exist

This is added in a different branch.

Co-authored-by: Leonardo Lopes de Albuquerque <leonardo.albuquerque@automattic.com>
Co-authored-by: Rasmy Nguyen <raz@automattic.com>