Replace unnecessary throws with empty strings when input is invalid (#2374)
* Replace unnecessary throws with empty strings when input is invalid

Before, if the key input was invalid an error was thrown. In some cases,
the error was not caught and was throwing fatal exceptions. Now, an
empty string is returned when a key input is invalid. This follows the
same pattern that exists in the validators of:
'woocommerce/includes/abstracts/abstract-wc-settings-api.php'

* Update changelog: 'Replace unnecessary throws with empty string when...'