Fixes math operation on empty string
`wc_parse_relative_date_option` will return an array. For that reason, it will cause to skip the validation because it will never be empty.

Later below, `$retention[‘number’]` is attempted unsuccessfully because `$retention[‘number’]` will be an empty string when the “Retain Stripe Data” option is not defined, causing an error and stopping the data removal.