Hide refund button for non-captured orders
Currently, when merchants have the "capture later" setting enabled, a
refund button is displayed for new orders which are authorized but not
captured. However, this makes no sense because there's no such thing as
refunding a charge that has not been made yet.

The current behavior is that if they do ANY refund, full or partial, it
will simply cancel the order and release the pre-authorization, but the
UI does a poor job of indicating that and the behavior is inherently
unintuitive.

This change simply checks if the payment for the order has been
authorized but not charged, and if so, hides the refund button and
replaces it with a disabled "Refunding unavailable" button which
includes a tooltip which explains why.

Fixes #2338.