Fix race condition when paying via UPE
Both the webhook processing and the UPE redirect processing call the
`process_response` function which marks the payment as completed. This
will cause the UPE redirect processing to fail due to the order status
if the webhook is processed before the customer is redirected, e.g.
when the customer is on a slow network connection.

This commit fixes that by adding metadata to the order that the webhook
controller can check for before processing the webhook.