Set transaction URL for Stripe APMs
When viewing order details, we want the charge ID to link to the Stripe dashboard.
We are already doing this for CCs, by overriding get_transaction_url() in
WC_Stripe_Payment_Gateway. We want to do this in WC_Stripe_UPE_Payment_Method as
well, to have the same behavior for APMs.