Fix error when discarding changes to the display order
When changes to the payment method order list are discarded via the
cancel button, it leads to an empty settings screen. This happens
because it tries to find the Link payment method in the mapping list but
fails and throw an error. We fix it by returning early if the Icon and
label fields for a payment method are not found on the mapping.